server/requirements.txt
fastapi==0.116.1
httpx==0.28.1
pydantic==1.10.22
python-dotenv==1.1.1
pytube==15.0.0
Requests==2.32.4
supabase==2.16.0
uvicorn==0.35.0
youtube_transcript_api==1.1.1
stripe==12.3.0